backend/pyproject.toml
[project]
name = "chartora-backend"
version = "0.1.0"
description = "Chartora - Quantum Computing Company Leaderboard Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "httpx>=0.28.0",
    "yfinance>=0.2.36",
    "sentry-sdk[fastapi]>=2.0.0",
    "stripe>=8.0.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.1.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short --strict-markers"
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
]

[tool.ruff]
target-version = "py311"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
ignore_missing_imports = true
plugins = []

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "src.infrastructure.database"
disallow_subclassing_any = false

[[tool.mypy.overrides]]
module = "src.adapters.api.schemas"
disallow_subclassing_any = false

[[tool.mypy.overrides]]
module = "yfinance.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sentry_sdk.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "stripe.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "jwt.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "bcrypt.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "src.adapters.api.routers.payments"
disallow_untyped_calls = false

[[tool.mypy.overrides]]
module = "src.adapters.api.middleware"
disallow_subclassing_any = false

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 80
